Moved basecoin into examples
