+++ examples/basecoin/cmd/basecli/LIGHT_NODE.md
