+++ examples/basecoin/cmd/basecli/README.md
