+++ examples/basecoin/cmd/basecli/main.go
