+++ examples/basecoin/cmd/basecoin/commands/init.go
