+++ examples/basecoin/cmd/basecoin/commands/key.go
