+++ examples/basecoin/cmd/basecoin/commands/reset.go
