+++ examples/basecoin/cmd/basecoin/commands/root.go
