+++ examples/basecoin/tests/cli/init.sh
@@ -105,4 +105,8 @@ checkDir() {
 
 # load and run these tests with shunit2!
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )" #get this files directory
-. $DIR/shunit2
+
+# TODO: how to handle this if we are not in the same directory
+CLI_DIR=${DIR}/../../../../tests/cli
+
+. $CLI_DIR/shunit2
