+++ examples/basecoin/tests/cli/rest.sh
@@ -151,5 +151,9 @@ test04CreateRoleInvalid() {
 
 # Load common then run these tests with shunit2!
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )" #get this files directory
-. $DIR/common.sh
-. $DIR/shunit2
+
+# TODO: how to handle this if we are not in the same directory
+CLI_DIR=${DIR}/../../../../tests/cli
+
+. $CLI_DIR/common.sh
+. $CLI_DIR/shunit2
