+++ examples/basecoin/tests/cli/restart.sh
@@ -78,6 +78,10 @@ test01OnRestart() {
 
 # Load common then run these tests with shunit2!
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )" #get this files directory
-. $DIR/common.sh
-. $DIR/shunit2
+
+# TODO: how to handle this if we are not in the same directory
+CLI_DIR=${DIR}/../../../../tests/cli
+
+. $CLI_DIR/common.sh
+. $CLI_DIR/shunit2
