+++ examples/basecoin/tests/cli/roles.sh
@@ -90,5 +90,9 @@ test03SendMultiFromRole() {
 
 # Load common then run these tests with shunit2!
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )" #get this files directory
-. $DIR/common.sh
-. $DIR/shunit2
+
+# TODO: how to handle this if we are not in the same directory
+CLI_DIR=${DIR}/../../../../tests/cli
+
+. $CLI_DIR/common.sh
+. $CLI_DIR/shunit2
